backend/requirements.txt
flask==2.3.3
werkzeug==2.3.7
flask-cors==4.0.0
nltk==3.8.1
spacy==3.7.2
gunicorn==21.2.0
python-dotenv==1.0.0
python-docx==0.8.11
PyPDF2==3.0.1
requests==2.31.0
beautifulsoup4==4.12.2